+++ tests/performance/3DLighting/src/main/java/attenuation/CameraScene3D.java
@@ -23,7 +23,7 @@
  * questions.
  */
 
-package lighting3D;
+package attenuation;
 
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.DoubleProperty;
