+++ tests/performance/3DLighting/src/main/java/attenuation/Controls.java
@@ -23,7 +23,7 @@
  * questions.
  */
 
-package lighting3D;
+package attenuation;
 
 import java.util.List;
