+++ tests/performance/3DLighting/src/main/java/attenuation/MaterialControls.java
@@ -1,4 +1,4 @@
-package lighting3D;
+package attenuation;
 
 import java.util.Optional;
